requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiomysql
aiosqlite
alembic
python-dotenv
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
google-genai
pytest
pytest-asyncio
httpx
python-multipart
greenlet
aiosqlite
fastapi
google-genai
greenlet
httpx
passlib
pydantic-settings
python-jose
python-multipart
sqlalchemy
uvicorn
